Common: A little hotfix

+++ src/Edde/Filter/IntFilter.php
@@ -7,11 +7,11 @@
 	class IntFilter extends AbstractFilter {
 		/** @inheritdoc */
 		public function input($value, ?array $options = null) {
-			return filter_var($value, FILTER_VALIDATE_INT);
+			return (int)filter_var($value, FILTER_VALIDATE_INT);
 		}
 
 		/** @inheritdoc */
 		public function output($value, ?array $options = null) {
-			return filter_var($value, FILTER_VALIDATE_INT);
+			return (int)filter_var($value, FILTER_VALIDATE_INT);
 		}
 	}
